ai-language-tutor/v4/backend/models/curriculum.py
from sqlalchemy import Column, Integer, String, JSON
from sqlalchemy.ext.declarative import declarative_base

Base = declarative_base()

class Curriculum(Base):
    __tablename__ = "curriculums"
    id = Column(Integer, primary_key=True, index=True)
    title = Column(String, nullable=False)
    metadata = Column(JSON) 