ai-language-tutor/v4/backend/utils/supabase_client.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv
import logging

# Load environment variables
load_dotenv()

# Get Supabase configuration
SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

# Validate configuration
if not SUPABASE_URL:
    raise ValueError("SUPABASE_URL environment variable is required")

if not SUPABASE_KEY:
    raise ValueError("SUPABASE_KEY environment variable is required")

# Create Supabase client
try:
    supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
    logging.info("Supabase client initialized successfully")
except Exception as e:
    logging.error(f"Failed to initialize Supabase client: {e}")
    raise

# Export for use in other modules
__all__ = ["supabase"] 